requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
aiohttp==3.8.6
Pillow==10.0.1
aiomysql==0.2.0